requirements.txt
torch
torchtext
numpy
scikit-learn
pyyaml